rename .hpp to .h (for R CMD check)
